refactor: expose kernel APIs needed by delta_kernel_default_engine
