+++ kernel/src/engine/plans/parquet.rs
@@ -64,23 +64,29 @@ mod tests {
     use std::sync::Arc;
 
     use tempfile::tempdir;
+    use url::Url;
 
     use super::PlanBasedParquetHandler;
     use crate::arrow::array::{Array, Int64Array, RecordBatch};
     use crate::engine::arrow_conversion::TryIntoKernel as _;
     use crate::engine::arrow_data::ArrowEngineData;
     use crate::engine::sync::plan::SyncPlanExecutor;
-    use crate::engine::tests::{
-        file_meta_for, test_parquet_handler_footer_errors_on_missing_file,
-        test_parquet_handler_footer_preserves_field_ids,
-        test_parquet_handler_reads_file_with_arrow_schema, test_parquet_handler_reads_footer,
-    };
     use crate::parquet::arrow::arrow_writer::ArrowWriter;
     use crate::schema::SchemaRef;
     use crate::{FileMeta, ParquetHandler as _};
 
     fn make_handler() -> PlanBasedParquetHandler {
         PlanBasedParquetHandler::new(Arc::new(SyncPlanExecutor::new()))
+    }
+
+    fn file_meta_for(path: &Path) -> FileMeta {
+        let url = Url::from_file_path(path).unwrap();
+        let size = std::fs::metadata(path).unwrap().len();
+        FileMeta {
+            location: url,
+            last_modified: 0,
+            size,
+        }
     }
 
     fn make_test_parquet_file(path: &Path, batch: &RecordBatch) -> (FileMeta, SchemaRef) {
@@ -127,25 +133,5 @@ mod tests {
                 .values(),
             &[10, 20, 30]
         );
-    }
-
-    #[test]
-    fn test_parquet_handler_reads_contract() {
-        test_parquet_handler_reads_file_with_arrow_schema(&make_handler());
-    }
-
-    #[test]
-    fn test_read_parquet_footer_contract() {
-        test_parquet_handler_reads_footer(&make_handler());
-    }
-
-    #[test]
-    fn test_read_parquet_footer_missing_file_contract() {
-        test_parquet_handler_footer_errors_on_missing_file(&make_handler());
-    }
-
-    #[test]
-    fn test_read_parquet_footer_preserves_field_ids_contract() {
-        test_parquet_handler_footer_preserves_field_ids(&make_handler());
     }
 }
