+++ test-utils/Cargo.toml
@@ -15,14 +15,17 @@ release = false
 [features]
 default = ["arrow"]
 arrow = ["arrow-58"]
-arrow-58 = ["delta_kernel/arrow-58"]
-arrow-57 = ["delta_kernel/arrow-57"]
+arrow-58 = ["delta_kernel/arrow-58", "delta_kernel_default_engine/arrow-58"]
+arrow-57 = ["delta_kernel/arrow-57", "delta_kernel_default_engine/arrow-57"]
 
 [dependencies]
 # NOTE: We MUST build with SOME arrow version (or compilation would fail). So we pull in `arrow-57`,
 # which ensures `cargo build` succeeds. But we ALSO default to `arrow`, which activates `arrow-58`
 # and overrides `arrow-57`.
-delta_kernel = { path = "../kernel", features = [ "default-engine-rustls", "arrow-57", "internal-api", "prettyprint", "test-utils" ] }
+delta_kernel = { path = "../kernel", features = [ "arrow-57", "internal-api", "prettyprint", "test-utils" ] }
+# Re-exported so kernel's integration tests (which cannot dev-dep default-engine directly due to
+# the cycle through this crate) can reach it via `test_utils::delta_kernel_default_engine::*`.
+delta_kernel_default_engine = { path = "../default-engine", default-features = false, features = ["arrow-57", "rustls"] }
 itertools = "0.14.0"
 serde_json = "1.0.142"
 tar = "0.4"
