+++ test-utils/src/engine_contract.rs
@@ -1,43 +1,43 @@
 //! Shared contract tests for engine handler traits ([`JsonHandler`], [`ParquetHandler`]).
 //!
 //! Each function here tests a specific piece of the handler contract using only the public
-//! trait API plus [`GetData`]/[`RowVisitor`] for result inspection — no engine-specific
+//! trait API plus [`GetData`]/[`RowVisitor`] for result inspection -- no engine-specific
 //! downcasting. Individual engine test modules call these to verify their implementation
 //! satisfies the contract, then add any engine-specific assertions (e.g. Arrow encoding
 //! details) in their own tests.
 
 use std::collections::HashMap;
-use std::path::PathBuf;
 use std::sync::{Arc, LazyLock};
 
+use delta_kernel::arrow::array::{Array, Int64Array, RecordBatch, StringArray};
+use delta_kernel::arrow::datatypes::{DataType as ArrowDataType, Field, Schema as ArrowSchema};
+use delta_kernel::engine::arrow_conversion::TryIntoKernel as _;
+use delta_kernel::engine::arrow_data::ArrowEngineData;
+use delta_kernel::engine_data::{FilteredEngineData, GetData, RowVisitor, TypedGetData as _};
+use delta_kernel::expressions::{column_name, ColumnName};
+use delta_kernel::object_store::path::Path;
+use delta_kernel::parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
+use delta_kernel::parquet::arrow::arrow_writer::ArrowWriter;
+use delta_kernel::parquet::arrow::{ARROW_SCHEMA_META_KEY, PARQUET_FIELD_ID_META_KEY};
+use delta_kernel::schema::{
+    ColumnMetadataKey, DataType, MetadataColumnSpec, PrimitiveType, SchemaRef, StructField,
+    StructType,
+};
+use delta_kernel::{DeltaResult, Engine, EngineData, FileMeta, JsonHandler, ParquetHandler};
 use itertools::Itertools;
 use tempfile::{tempdir, NamedTempFile};
-use test_utils::delta_path_for_version;
 use url::Url;
 
-use crate::arrow::array::{Array, Int64Array, RecordBatch, StringArray};
-use crate::arrow::datatypes::{DataType as ArrowDataType, Field, Schema as ArrowSchema};
-use crate::engine::arrow_conversion::TryIntoKernel as _;
-use crate::engine::arrow_data::ArrowEngineData;
-use crate::engine_data::{FilteredEngineData, GetData, RowVisitor, TypedGetData as _};
-use crate::object_store::path::Path;
-use crate::parquet::arrow::arrow_reader::ParquetRecordBatchReaderBuilder;
-use crate::parquet::arrow::arrow_writer::ArrowWriter;
-use crate::parquet::arrow::{ARROW_SCHEMA_META_KEY, PARQUET_FIELD_ID_META_KEY};
-use crate::schema::{
-    column_name, ColumnMetadataKey, ColumnName, ColumnNamesAndTypes, DataType, MetadataColumnSpec,
-    StructField, StructType,
-};
-use crate::{DeltaResult, Engine, EngineData, FileMeta, JsonHandler, ParquetHandler};
+use crate::delta_path_for_version;
 
-// ---------------------------------------------------------------------------
+// ===========================================================================
 // Shared file-setup helpers
-// ---------------------------------------------------------------------------
+// ===========================================================================
 
 /// Writes `lines` as newline-delimited JSON to a [`NamedTempFile`] and returns the file
 /// together with a [`FileMeta`] pointing at it. The temp file must be kept alive for as
 /// long as the `FileMeta` is in use.
-pub(crate) fn make_temp_json_file(lines: &[&str]) -> (NamedTempFile, FileMeta) {
+pub fn make_temp_json_file(lines: &[&str]) -> (NamedTempFile, FileMeta) {
     let mut temp_file = NamedTempFile::new().expect("Failed to create temp file");
     for line in lines {
         use std::io::Write as _;
@@ -57,7 +57,7 @@ pub(crate) fn make_temp_json_file(lines: &[&str]) -> (NamedTempFile, FileMeta) {
 }
 
 /// Builds a [`FileMeta`] for a local file path, reading the actual size from the filesystem.
-pub(crate) fn file_meta_for(path: &std::path::Path) -> FileMeta {
+pub fn file_meta_for(path: &std::path::Path) -> FileMeta {
     let url = Url::from_file_path(path).unwrap();
     let size = std::fs::metadata(path).unwrap().len();
     FileMeta {
@@ -67,14 +67,14 @@ pub(crate) fn file_meta_for(path: &std::path::Path) -> FileMeta {
     }
 }
 
-// ---------------------------------------------------------------------------
+// ===========================================================================
 // JsonHandler contract tests
-// ---------------------------------------------------------------------------
+// ===========================================================================
 
 /// Contract: any [`JsonHandler`] that receives a schema with a [`MetadataColumnSpec::FilePath`]
 /// column must populate it with the file URL for every row, readable via [`GetData`] without
 /// any Arrow downcasting.
-pub(crate) fn test_json_handler_file_path_contract(handler: &dyn JsonHandler) {
+pub fn test_json_handler_file_path_contract(handler: &dyn JsonHandler) {
     let (_temp, file_meta) = make_temp_json_file(&[r#"{"x": 1}"#, r#"{"x": 2}"#]);
     let expected_url = file_meta.location.to_string();
 
@@ -98,9 +98,9 @@ pub(crate) fn test_json_handler_file_path_contract(handler: &dyn JsonHandler) {
     }
     impl RowVisitor for FilePathCollector {
         fn selected_column_names_and_types(&self) -> (&'static [ColumnName], &'static [DataType]) {
-            static NAT: LazyLock<ColumnNamesAndTypes> =
-                LazyLock::new(|| (vec![column_name!("_file")], vec![DataType::STRING]).into());
-            NAT.as_ref()
+            static NAMES: LazyLock<Vec<ColumnName>> = LazyLock::new(|| vec![column_name!("_file")]);
+            static TYPES: LazyLock<Vec<DataType>> = LazyLock::new(|| vec![DataType::STRING]);
+            (&NAMES, &TYPES)
         }
         fn visit<'a>(
             &mut self,
@@ -124,25 +124,96 @@ pub(crate) fn test_json_handler_file_path_contract(handler: &dyn JsonHandler) {
     );
 }
 
-// ---------------------------------------------------------------------------
+// ===========================================================================
 // ParquetHandler contract tests
-// ---------------------------------------------------------------------------
+// ===========================================================================
 
-/// Contract: [`ParquetHandler::read_parquet_footer`] must correctly parse the schema
-/// from a real Delta checkpoint file.
-pub(crate) fn test_parquet_handler_reads_footer(handler: &dyn ParquetHandler) {
-    let path = std::fs::canonicalize(PathBuf::from(
-        "./tests/data/with_checkpoint_no_last_checkpoint/_delta_log/00000000000000000002.checkpoint.parquet",
-    ))
-    .unwrap();
+/// Contract: [`ParquetHandler::read_parquet_footer`] must correctly parse the schema from a
+/// real Delta checkpoint file. `checkpoint_path` should point at a V1 checkpoint parquet file
+/// (kernel ships a fixture under `kernel/tests/data/`).
+pub fn test_parquet_handler_reads_footer(
+    handler: &dyn ParquetHandler,
+    checkpoint_path: &std::path::Path,
+) {
+    let path = std::fs::canonicalize(checkpoint_path).unwrap();
     let file_meta = file_meta_for(&path);
     let footer = handler.read_parquet_footer(&file_meta).unwrap();
-    crate::utils::test_utils::validate_checkpoint_schema(&footer.schema);
+    validate_checkpoint_schema(&footer.schema);
+}
+
+/// Validates that a schema has the expected checkpoint structure with top-level action fields
+/// and proper nested types for add, metaData, and protocol actions.
+fn validate_checkpoint_schema(schema: &SchemaRef) {
+    fn get_field(struct_type: &StructType, name: &str) -> StructField {
+        struct_type
+            .fields()
+            .find(|f| f.name() == name)
+            .unwrap_or_else(|| panic!("Field '{name}' not found"))
+            .clone()
+    }
+
+    for field_name in ["txn", "add", "remove", "metaData", "protocol"] {
+        let field = get_field(schema, field_name);
+        assert!(
+            matches!(field.data_type(), DataType::Struct(_)),
+            "Field '{field_name}' should be a struct type"
+        );
+    }
+
+    let add_field = get_field(schema, "add");
+    let add_struct = match add_field.data_type() {
+        DataType::Struct(s) => s,
+        _ => panic!("'add' should be a struct"),
+    };
+    assert_eq!(
+        get_field(add_struct, "path").data_type(),
+        &DataType::Primitive(PrimitiveType::String)
+    );
+    assert_eq!(
+        get_field(add_struct, "size").data_type(),
+        &DataType::Primitive(PrimitiveType::Long)
+    );
+    assert!(
+        matches!(
+            get_field(add_struct, "partitionValues").data_type(),
+            DataType::Map(_)
+        ),
+        "'partitionValues' should be a map type"
+    );
+
+    let metadata_field = get_field(schema, "metaData");
+    let metadata_struct = match metadata_field.data_type() {
+        DataType::Struct(s) => s,
+        _ => panic!("'metaData' should be a struct"),
+    };
+    let format_field = get_field(metadata_struct, "format");
+    let format_struct = match format_field.data_type() {
+        DataType::Struct(s) => s,
+        _ => panic!("'format' should be a struct"),
+    };
+    assert_eq!(
+        get_field(format_struct, "provider").data_type(),
+        &DataType::Primitive(PrimitiveType::String)
+    );
+
+    let protocol_field = get_field(schema, "protocol");
+    let protocol_struct = match protocol_field.data_type() {
+        DataType::Struct(s) => s,
+        _ => panic!("'protocol' should be a struct"),
+    };
+    assert_eq!(
+        get_field(protocol_struct, "minReaderVersion").data_type(),
+        &DataType::Primitive(PrimitiveType::Integer)
+    );
+    assert_eq!(
+        get_field(protocol_struct, "minWriterVersion").data_type(),
+        &DataType::Primitive(PrimitiveType::Integer)
+    );
 }
 
 /// Contract: [`ParquetHandler::read_parquet_footer`] must return an error for a
 /// non-existent file.
-pub(crate) fn test_parquet_handler_footer_errors_on_missing_file(handler: &dyn ParquetHandler) {
+pub fn test_parquet_handler_footer_errors_on_missing_file(handler: &dyn ParquetHandler) {
     let mut temp_path = std::env::temp_dir();
     temp_path.push("non_existent_kernel_test_file.parquet");
     let file_meta = FileMeta {
@@ -158,7 +229,7 @@ pub(crate) fn test_parquet_handler_footer_errors_on_missing_file(handler: &dyn P
 
 /// Contract: [`ParquetHandler::read_parquet_footer`] must preserve Arrow field IDs,
 /// accessible via [`ColumnMetadataKey::ParquetFieldId`].
-pub(crate) fn test_parquet_handler_footer_preserves_field_ids(handler: &dyn ParquetHandler) {
+pub fn test_parquet_handler_footer_preserves_field_ids(handler: &dyn ParquetHandler) {
     let make_field_with_id = |name: &str, ty: ArrowDataType, nullable: bool, id: &str| {
         Field::new(name, ty, nullable).with_metadata(HashMap::from([(
             PARQUET_FIELD_ID_META_KEY.to_string(),
@@ -205,7 +276,7 @@ pub(crate) fn test_parquet_handler_footer_preserves_field_ids(handler: &dyn Parq
 /// Contract: [`ParquetHandler::write_parquet_file`] always overwrites an existing file.
 /// Writes `[1, 2, 3]`, then overwrites with `[10, 20]`, and verifies only `[10, 20]`
 /// is present.
-pub(crate) fn test_parquet_handler_write_always_overwrites(handler: &dyn ParquetHandler) {
+pub fn test_parquet_handler_write_always_overwrites(handler: &dyn ParquetHandler) {
     let temp_dir = tempdir().unwrap();
     let file_path = temp_dir.path().join("overwrite_test.parquet");
     let url = Url::from_file_path(&file_path).unwrap();
@@ -267,7 +338,7 @@ pub(crate) fn test_parquet_handler_write_always_overwrites(handler: &dyn Parquet
 
 /// Contract: [`ParquetHandler::write_parquet_file`] must not embed the Arrow IPC schema
 /// (`ARROW:schema`) in the file's key/value metadata.
-pub(crate) fn test_parquet_handler_write_omits_arrow_schema(handler: &dyn ParquetHandler) {
+pub fn test_parquet_handler_write_omits_arrow_schema(handler: &dyn ParquetHandler) {
     let temp_dir = tempdir().unwrap();
     let file_path = temp_dir.path().join("no_arrow_schema.parquet");
     let url = Url::from_file_path(&file_path).unwrap();
@@ -298,7 +369,7 @@ pub(crate) fn test_parquet_handler_write_omits_arrow_schema(handler: &dyn Parque
 /// Contract: [`ParquetHandler::read_parquet_files`] must successfully read a file whose
 /// metadata contains an embedded Arrow IPC schema (e.g. one produced by [`ArrowWriter`]
 /// without explicit options to skip it).
-pub(crate) fn test_parquet_handler_reads_file_with_arrow_schema(handler: &dyn ParquetHandler) {
+pub fn test_parquet_handler_reads_file_with_arrow_schema(handler: &dyn ParquetHandler) {
     let temp_dir = tempdir().unwrap();
     let file_path = temp_dir.path().join("with_arrow_schema.parquet");
 
@@ -341,11 +412,11 @@ pub(crate) fn test_parquet_handler_reads_file_with_arrow_schema(handler: &dyn Pa
     );
 }
 
-// ---------------------------------------------------------------------------
+// ===========================================================================
 // Storage / Engine helpers (used by the engine-level tests)
-// ---------------------------------------------------------------------------
+// ===========================================================================
 
-pub(crate) fn test_arrow_engine(engine: &dyn Engine, base_url: &Url) {
+pub fn test_arrow_engine(engine: &dyn Engine, base_url: &Url) {
     test_list_from_should_sort_and_filter(engine, base_url, get_arrow_data);
 }
